requirements.txt
Flask==0.10.1
Pillow==2.5.0
Flask-Script
riak
python-magic==0.4.6
requests